requirements.txt
pyrogram
tgcrypto
feedparser
google-generativeai